Fixing type size error in DMProc new writer for prompt and disassembler fix.

+++ DMCompiler/DM/Optimizer/AnnotatedByteCodeWriter.cs
@@ -125,7 +125,7 @@ namespace DMCompiler.DM.Optimizer {
             _requiredArgs.Pop();
 
             _annotatedBytecode[^1].AddArg(new AnnotatedBytecodeType(type, location));
-            Write((byte)type);
+            Write((int)type);
 
         }
